% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishBearingFlow.r
\name{calcFishBearingFlow}
\alias{calcFishBearingFlow}
\title{Redistribute flows from ResSim according to calculated FPS flow}
\usage{
calcFishBearingFlow(fish_postDPE, param_list, verbose)
}
\arguments{
\item{fish_postDPE}{Dataframe containing daily estimates of various
hydrological data (e.g., ResSim estimates of flow) in addition to the prop.
of annually approaching fish who are actually in the dam.
This dataframe must include a column, \code{approaching_daily_postDPE}, the daily
proportion of fish that pass the dam (after accounting for dam passage
efficiency)}

\item{param_list}{A named list of dam operation and biological parameters required to
run FBW.
This list is typically created by reading in values from a template using the \code{loadFromTemplate} or
\code{loadFromWorkbook} functions. \if{html}{\out{<br>}} The list should include at least the following entries:
\enumerate{
\item \code{alt_desc}, with named entry "collector", the fish passage structure name.
\item \code{route_dpe}, a dataframe including columns \code{elev}, \code{baseline_dpe}, and any
number of other columns to the right of these that can be used to look up DPE
at various pool elevations).
\item \code{fps_max_elev}: a single numeric value, the maximum pool elevation at
which the fish passage structure can operate).
\item \code{fps_bottom_elev}: a single numeric value, the minimum pool elevation (in
feet) at which the fish passage structure can operate.
\item \code{dpe_x_position}: a single numeric value referencing which column of the
DPE lookup table (provided in \code{param_list}). Used to determine how many
columns of \code{route_dpe} AFTER the baseline should be selected?
}}

\item{verbose}{(for checking of outputs) If \code{FALSE}, the default value, the
returned dataframe will not include intermediate calculations (e.g., monthly
flow rates). It may be helpful for debugging purposes to retain the columns
that contain these intermediate steps. Setting \code{verbose = TRUE} will return
all columns.}
}
\value{
A dataframe with all of the columns of the input table,
\code{fish_postDPE}, with the addition of some number of columns (depending on the
\code{verbose} argument). At least the flow rate through the fish passage
structure, \code{FPS_flow}, and the proportion of fish passing through each
outlet in the dam is returned. Optionally, if \code{verbose = TRUE},
intermediate variables like the proportion of flow through each outlet, route
effectiveness, and fish-bearing flow are output.
}
\description{
First distributes fish bearing flow through available outlets in
a dam, including fish passage structure calculations. Then, using the
proportion of fish-bearing flow, calculates the distribution of fish passing
the dam according to route effectiveness and fish-bearing flows.
}
