% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTempSplit.r
\name{calcTempSplit}
\alias{calcTempSplit}
\title{Calculate the distribution of flow between the temperature control tower and
the fish passage structure in cases where there is temperature control (FSS)}
\usage{
calcTempSplit(ressim_data, water_year_types, temp_dist_df)
}
\arguments{
\item{ressim_data}{A dataframe that includes at least Date (in datetime
format) and information on daily Outflow (cfs), temperature splits}

\item{water_year_types}{A dataframe containing "year" and "watertype", one
of "abundant", "adequate", "insufficient", and "deficient"}

\item{temp_dist_df}{a dataframe containing dated temperature splits (often
month-by-month) based on year types: "C/W" (cool/wet), "N" (normal), and
"H/D" (hot/dry)}
}
\value{
Dataframe of temperature splits in each day (a proportion, 0-1,
reflecting which dates have what percent of the outflow will be redirected
for temperature control operations). Flow through the fish passage structure
is calculated as outflow * (1 - proportion redirected to temperature control)
}
\description{
Calculate the distribution of flow between the temperature control tower and
the fish passage structure in cases where there is temperature control (FSS)
}
