% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributeFish_outlets.r
\name{distributeFish_outlets}
\alias{distributeFish_outlets}
\title{Calculate distribution of fish through outlets}
\usage{
distributeFish_outlets(fish_postDPE, param_list, verbose = FALSE)
}
\arguments{
\item{fish_postDPE}{Dataframe containing daily estimates of various
hydrological data (e.g., ResSim estimates of flow) in addition to the prop.
of annually approaching fish who are actually in the dam.
This dataframe must include a column, \code{approaching_daily_postDPE}, the daily
proportion of fish that pass the dam (after accounting for dam passage
efficiency)}

\item{param_list}{A list including at least the following named objects:
\enumerate{
\item \code{alt_desc}, with named entry "collector", the fish passage structure name;
\item \code{route_dpe}, a dataframe including columns \code{elev}, \code{baseline_dpe}, and any
number of other columns to the right of these that can be used to look up DPE
at various pool elevations);
\item \code{fps_max_elev}: a single numeric value, the maximum pool elevation at
which the fish passage structure can operate);
\item \code{fps_bottom_elev}: a single numeric value, the minimum pool elevation (in
feet) at which the fish passage structure can operate;
\item \code{dpe_column_name}: the name of the column in the \code{route_dpe} table that should
be used to inform elevation-specific DPE values. This can be one of \code{baseline_dpe}
or any other non-elevation columns in the \code{route_dpe} table. Typically, non-baseline
DPE values are used when a fish passage structure is in place. When a non-baseline DPE
column is used, it only applies if the pool elevation is within the operating
elevations of the passage structure.
}}

\item{verbose}{(Optional) Logical argument indicating whether to retain
intermediate columns in the output (proportion of spill in each outlet: \code{pB.spill},
\code{pB.turb}, \code{pB.RO}, and \code{pB.FPS}; fish-bearing flow through each
outlet: \code{B.spill}, \code{B.turb}, \code{B.RO}, and \code{B.FPS}; and route
effectiveness for each outlet: \code{RE.spill}, \code{RE.turb}, \code{RE.RO}, and
\code{RE.FPS}). Defaults to \code{FALSE}, in which case only the proportion of fish
through each outlet is returned in the output dataframe.}
}
\value{
A dataframe with all of the columns of the input table,
\code{fish_postDPE}, with the addition of some number of columns (depending on the
\code{verbose} argument). At least the flow rate through the fish passage
structure, \code{FPS_flow}, and the proportion of fish passing through each
outlet in the dam is returned. Optionally, if \code{verbose = TRUE},
intermediate variables like the proportion of flow through each outlet, route
effectiveness, and fish-bearing flow are output.
}
\description{
First distributes fish bearing flow through available outlets in
a dam, including fish passage structure calculations. Then, using the
proportion of fish-bearing flow, calculates the distribution of fish passing
the dam according to route effectiveness and fish-bearing flows.
}
