% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbwR-package.R
\docType{package}
\name{fbwR-package}
\alias{fbwR}
\alias{fbwR-package}
\title{fbwR: Fish Benefits Workbook in R: Estimating Downstream Fish Passage and Survival}
\description{
The Fish Benefits Workbook (FBW) is a simulation model that uses a series of calculations to estimate how dam structures, operations, and hydrology impact downstream survival of fish passing dams. It was developed from an Excel Workbook, the Fish Benefits Workbook, into R for the sake of expanding the model and to allow for greater use and dissemination of the model. The model is based on information about fish timing, dam operations from a hydrological model (ResSim, developed by the US Army Corps' Hydrological Engineering Center), and estimates of dam passage efficiency and survival from field studies. FBW simulates the distribution of flow through a dam's outlets routes fish through various outlets in a dam, including spillways, turbines, regulating outlets, and a variety of types of fish passage structures. Calculations are performed for each day within a hydrological period of record defined in ResSim. Results from FBW include estimated dam passage efficiency (the proportion of fish expected to successfully pass a dam if they are attempting passage), distribution of fish through available routes through the dam, and passage survival through each route.
}
\author{
\strong{Maintainer}: Mairin Deith \email{mairin.deith@gmail.com}

}
\keyword{internal}
