% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDPE.r
\name{fetchDPE}
\alias{fetchDPE}
\title{Look up dam passage efficiency (DPE) for a given set of fish passage
structures and pool elevation (and elevation limits)}
\usage{
fetchDPE(ressim, param_list)
}
\arguments{
\item{ressim}{Data frame including at least the following column names:
elev (pool elevation, one per day in the period of record);}

\item{param_list}{A list including at least the following named objects:
\code{alt_desc}, with named entry "collector", the fish passage structure name
\code{route_dpe}, a dataframe including columns \code{elev}, \code{baseline_dpe}, and any
number of other columns to the right of these that can be used to look up DPE
at various pool elevations)
\code{fps_max_elev}: a single numeric value, the maximum pool elevation at
which the fish passage structure can operate)
\code{fps_bottom_elev}: a single numeric value, the minimum pool elevation (in
feet) at which the fish passage structure can operate
\code{dpe_x_position}: a single numeric value referencing which column of the
DPE lookup table (provided in \code{param_list}). Used to determine how many
which column of \code{route_dpe} AFTER the baseline should be selected?}
}
\value{
A vector of DPE with one entry for each pool elevation provided.
}
\description{
Look up dam passage efficiency (DPE) for a given set of fish passage
structures and pool elevation (and elevation limits)
}
