% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromTemplate.r
\name{loadFromTemplate}
\alias{loadFromTemplate}
\title{Loads FBW data from a defined template spreadsheet in XLSX format.}
\usage{
loadFromTemplate(template_file)
}
\arguments{
\item{template_file}{File path to an Excel spreadsheet with standardized
inputs}
}
\description{
Loads FBW data from a defined template spreadsheet in XLSX format.
}
