% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromWorkbook.r
\name{loadFromWorkbook}
\alias{loadFromWorkbook}
\title{Load FBW parameters from an Excel workbook, requiring only the name of the
reservoir and the name of the quickset. If none are given, the reservoir
name and quickset name are read from the "Route Survival Model" landing page
of the FBW workbook.}
\usage{
loadFromWorkbook(fbw_excel, reservoir = NULL, quickset = NULL)
}
\arguments{
\item{fbw_excel}{Character string referencing the location of the Excel file
to be read. Parameters will be compiled from the ResvData and QuickSets
sheets.}

\item{reservoir}{(Optional) Name of the reservoir being modelled. Used to
lookup which entries of the ResvData and QuickSets should be loaded into FBW.
If not provided, \code{reservoir} is read from cell B6 of the Route Survival Model
sheet in \code{fbw_excel}.}

\item{quickset}{(Optional) Name of the quickset being modelled. Used to
lookup which entries of the QuickSets sheet should be loaded into FBW.
If not \code{reservoir} and \code{quickset} are not provided, \code{quickset} is read from
cell B7 of the Route Survival Model sheet in \code{fbw_excel}.}
}
\value{
A list of parameters required to run the FBW model in R.
}
\description{
Load FBW parameters from an Excel workbook, requiring only the name of the
reservoir and the name of the quickset. If none are given, the reservoir
name and quickset name are read from the "Route Survival Model" landing page
of the FBW workbook.
}
