% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResSim.r
\name{loadResSim}
\alias{loadResSim}
\title{Load data from an Excel ResSim file.}
\usage{
loadResSim(infile, wide = TRUE)
}
\arguments{
\item{infile}{A character string indicating the full file path to an
Excel workbook of HEC-ResSim flow outputs. Each of the following worksheets
must be present (\code{*} indicates a prefix that will be ignored when the file
is read in): \verb{*_ELEV}: elevation of the pool in feet; \verb{*_OUT}:
total outflow in cubic feet per second (CFS); \verb{*_PH}: flow through the
powerhouse turbines (total) in CFS; \verb{*_RO}: flow through the regulating
outlet in CFS; \verb{*_SPILL}: flow through the spillway in CFS. In each sheet,
there should be one column for each year in the period of record and one
row for each day in the year.}

\item{wide}{Are data in wide format (e.g., year across columns, one row for each
day in the 365-day year, Feb 29 in leap years are ignored)? Defaults to TRUE,
following standard input format of Res-Sim spreadsheets.}
}
\description{
Load data from an Excel ResSim file.
}
