% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromWorkbook.r
\name{loadResvData_FBWworkbook}
\alias{loadResvData_FBWworkbook}
\title{Load parameters required for FBW-R from a pre-filled Excel workbook using
two superceded functions: loadResvData_FBWworkbook and
loadQuickset_FBWworkbook}
\usage{
loadResvData_FBWworkbook(infile, resv_column)
}
\arguments{
\item{infile}{Character path to an Excel file that corresponds to a
parameterized FBW workbook}

\item{resv_column}{Character of the column in the Excel Workbook's "ResvData"
sheet that corresponds to the current dam.}
}
\description{
Load parameters required for FBW-R from a pre-filled Excel workbook using
two superceded functions: loadResvData_FBWworkbook and
loadQuickset_FBWworkbook
}
