% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_closest.r
\name{match_closest}
\alias{match_closest}
\title{Find the closet value to x in a table, given a provided tolerance value.
Taken from the R MALDIQuant library}
\usage{
match_closest(x, table, tolerance = Inf, nomatch = NA_integer_)
}
\arguments{
\item{x}{Value to look up for matching in the table}

\item{table}{Table of values with columns to search/lookup}

\item{tolerance}{Accepted tolerance between the value to look up and the
closest value in the lookup table. If distance between x and the closest
value in table is greater than tolerance, nomatch is returned. Defaults to
infinity, such that no tolerance is applied.}

\item{nomatch}{The value to return if the difference between x and the
closest value in the table is greater than the tolerance value}
}
\value{
Closest value in table to x.
}
\description{
Find the closet value to x in a table, given a provided tolerance value.
Taken from the R MALDIQuant library
}
