% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeFBW.r
\name{summarizeFBW}
\alias{summarizeFBW}
\title{Summarize daily FBW outputs into by-month and by-water-year-type summarize.}
\usage{
summarizeFBW(fish_passage_survival)
}
\arguments{
\item{fish_passage_survival}{A dataframe created by runFBW that includes
daily estimates of at least flow rates, survival rates, pool elevation,
and distribution of fish within the dam's routes. This can be created by
runFBW() if summarize is set to FALSE.}
}
\value{
A list containing two objects: a summary by month, and a summary
by water year type in the period of record, including stats::quantiles.
}
\description{
Summarize daily FBW outputs into by-month and by-water-year-type summarize.
}
