% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributeFishDaily.r
\name{distributeFishDaily}
\alias{distributeFishDaily}
\alias{distributeFish_daily}
\title{Distribute fish approaching a dam in a given month into daily approaching
subpopulations.}
\usage{
distributeFishDaily(ressim, param_list, verbose = FALSE)
}
\arguments{
\item{ressim}{A dataframe that includes at least Date (in datetime
format) and Outflow (cfs)}

\item{param_list}{A list including at least the following named objects:
\code{alt_desc}, with named entry "collector", the fish passage structure name
\code{route_dpe}, a dataframe including columns \code{elev}, \code{baseline_dpe}, and any
number of other columns to the right of these that can be used to look up DPE
at various pool elevations)
\code{fps_max_elev}: a single numeric value, the maximum pool elevation at
which the fish passage structure can operate)
\code{fps_bottom_elev}: a single numeric value, the minimum pool elevation (in
feet) at which the fish passage structure can operate
\code{dpe_x_position}: a single numeric value referencing which column of the
DPE lookup table (provided in \code{param_list}). Used to determine how many
which column of \code{route_dpe} AFTER the baseline should be selected?}

\item{verbose}{(for checking of outputs) If \code{FALSE}, the default value, the
returned dataframe will not include intermediate calculations (e.g., monthly
flow rates). It may be helpful for debugging purposes to retain the columns
that contain these intermediate steps. Setting \code{verbose = TRUE} will return
all columns.}
}
\value{
Dataframe matching ressim but with new columns appended:
"fry_DailyDistribution", "sub_DailyDistribution", "year_DailyDistribution"
containing the calculated proportion of fry, sub-yearlings, and yearlngs
approaching the dam on each day
}
\description{
Distribute fish approaching a dam in a given month into daily approaching
subpopulations.
}
