% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributeFlow_Survival_gates.r
\name{distributeFlow_Survival_gates}
\alias{distributeFlow_Survival_gates}
\title{Calculate distribution of flow through individual gates within each dam
outlet. This is useful in cases when a given dam outlet -- for example,
turbines -- have multiple sub-outlets through which a fish can pass the dam.
This function distributes flow between sub-outlets, and calculates survival
through those individual gates as a function of flow (or as a fixed value,
depending on parameterization).}
\usage{
distributeFlow_Survival_gates(fish_distributed_outlets, param_list)
}
\arguments{
\item{fish_distributed_outlets}{A list including at least the following named
objects:
route_dpe (a dataframe including columns \code{elev}, \code{baseline_dpe}, and any
number of other columns to the right of these that can be used to look up DPE
at various pool elevations).}

\item{param_list}{Parameter list created by \code{loadFromTemplate} and
\code{loadFromWorkbook} functions.}
}
\description{
Calculate distribution of flow through individual gates within each dam
outlet. This is useful in cases when a given dam outlet -- for example,
turbines -- have multiple sub-outlets through which a fish can pass the dam.
This function distributes flow between sub-outlets, and calculates survival
through those individual gates as a function of flow (or as a fixed value,
depending on parameterization).
}
