% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromTemplate.r
\name{loadFromTemplate}
\alias{loadFromTemplate}
\title{Loads FBW data from a defined template spreadsheet in XLSX format.}
\usage{
loadFromTemplate(
  template_file,
  year_override = FALSE,
  forced_year_range = NULL
)
}
\arguments{
\item{template_file}{File path to an Excel spreadsheet with standardized
inputs}

\item{year_override}{Boolean; should the years in the period of record be
overwritten by the years in \code{forced_year_range}? If TRUE, \code{forced_year_range}
should also be provided.}

\item{forced_year_range}{If \code{year_override} is TRUE, what years should be
imposed on the ResSim data?}
}
\description{
Loads FBW data from a defined template spreadsheet in XLSX format.
}
