% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFBW.r
\name{runFBW}
\alias{runFBW}
\title{Loads FBW data from a defined template spreadsheet in XLSX format.}
\usage{
runFBW(ressim_file, template_file, ressim_wide = TRUE, summarize = FALSE)
}
\arguments{
\item{ressim_file}{File path to an Excel spreadsheet with ResSim inputs}

\item{template_file}{File path to an Excel spreadsheet with standardized
inputs}

\item{ressim_wide}{Are ResSim data in typical wide format? Defaults to TRUE,
which assumes that each column in the dataframe contains data for a single
year, with each row representing a single day in the 365-day year.}

\item{summarize}{Should the daily outputs be summarized into average monthly
survival estimates? This summarizes across years within the period of record
in the ResSim input file.}
}
\description{
Loads FBW data from a defined template spreadsheet in XLSX format.
}
