% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFBW.r
\name{runFBW}
\alias{runFBW}
\title{Loads FBW data from a defined template spreadsheet in XLSX format.}
\usage{
runFBW(
  template_file = NULL,
  param_list = NULL,
  ressim_path = NULL,
  ressim = NULL,
  ressim_wide = TRUE,
  summarize = FALSE
)
}
\arguments{
\item{param_list}{Named list of FBW parameters created when the template
is read. This allows parameters to be modified or created manually without
having to create a separate template file. One of \code{template_path} or
\code{param_list} must be provided.}

\item{ressim_path}{File path to an Excel spreadsheet with ResSim inputs.
One of \code{ressim_path} or \code{ressim_file} must be provided.}

\item{ressim}{A loaded file (e.g., using the \code{loadResSim} function)
which may be provided instead of \code{ressim_path}. One of \code{ressim_path} or
\code{ressim_file} must be provided.}

\item{ressim_wide}{Are ResSim data in typical wide format? Defaults to TRUE,
which assumes that each column in the dataframe contains data for a single
year, with each row representing a single day in the 365-day year.}

\item{summarize}{Should the daily outputs be summarized into average monthly
survival estimates? This summarizes across years within the period of record
in the ResSim input file.}

\item{template_path}{File path to an Excel spreadsheet with standardized
inputs. One of \code{template_path} or \code{param_list} must be provided; if using
a template, it is loaded and translated into a parameter list like
\code{param_list}.}
}
\description{
Loads FBW data from a defined template spreadsheet in XLSX format.
}
